assertions

+++ TesteDeSoftwareDevIo/TestesDeUnidade/Demo/Calculadora.cs
@@ -5,7 +5,7 @@
         public double Somar(double v1, double v2) =>
             v1 + v2;
 
-        public double Dividir(double v1, double v2) =>
+        public double Dividir(int v1, int v2) =>
             v1 / v2;
     }
 }
